requirements.txt
googletrans
tqdm
pandas
nltk
transformers
Flask
python-docx